factor code into a test package
