+++ test/ReferenceTests/Project.toml
@@ -1,0 +1,25 @@
+name = "ReferenceTests"
+uuid = "d37af2e0-5618-4e00-9939-d430db56ee94"
+authors = ["Simon <sdanisch@protonmail.com> and contributors"]
+version = "0.1.0"
+
+[deps]
+AbstractPlotting = "537997a7-5e4e-5d89-9595-2241ea00577e"
+ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
+FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+FixedPointNumbers = "53c48c17-4a7d-5ca2-90c5-79b7896eea93"
+GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
+Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+MeshIO = "7269a6da-0436-5bbc-96c2-40638cbb6118"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+ghr_jll = "07c12ed4-43bc-5495-8a2a-d5838ef8d533"
+
+[compat]
+julia = "1.3"
