+++ test/reference_tests.jl
@@ -1,0 +1,6 @@
+Pkg.develop(path="./ReferenceTests")
+using ReferenceTests
+@testset "reference image tests" begin
+    ReferenceTests.run_tests()
+end
+
